Measure reprojection error
The reprojection error doesn't really seem to be that good of a measure
of uncertainty, at least for my video data, since the more markers you
see, the higher the errors get. Combining the number of visible markers
with the reprojection error in some way could help fix this. However,
the reprojection error is at least some kind of a measure that could be
used for measuring uncertainty.

To have something to show to the user, show the reprojection errors
using colors in the 3D visualization window. Green means low error and
red means high error.

References: https://github.com/ttthi/bananas/issues/46